# frontend/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PromptRepo API",
    "description": "Backend API for PromptRepo application with GitHub OAuth",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/prompts/": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get prompt",
        "description": "Get a specific prompt by repository name and file path. Checks user permissions based on hosting type.",
        "operationId": "get_prompt_api_v0_prompts__get",
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          },
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_PromptMeta_"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Prompt Not Found",
                  "detail": "Prompt with ID 'xxx' not found or access denied"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to retrieve prompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Create prompt",
        "description": "Create a new prompt. The prompt will be saved to the specified repository and file path.",
        "operationId": "create_prompt_api_v0_prompts__post",
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          },
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_PromptMeta_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid prompt data",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/validation-failed",
                  "title": "Validation Error",
                  "detail": "Invalid prompt data"
                }
              }
            }
          },
          "409": {
            "description": "Prompt file already exists",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/conflict",
                  "title": "Conflict",
                  "detail": "Prompt file already exists at the specified path"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to create prompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update prompt",
        "description": "Update an existing prompt. All fields in the update are optional.",
        "operationId": "update_prompt_api_v0_prompts__put",
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          },
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptDataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_PromptMeta_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid prompt data",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/validation-failed",
                  "title": "Validation Error",
                  "detail": "Invalid prompt data"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Prompt Not Found",
                  "detail": "Prompt with ID 'xxx' not found or access denied"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to update prompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prompts"
        ],
        "summary": "Delete prompt",
        "description": "Delete a prompt. Removes the prompt file from the repository.",
        "operationId": "delete_prompt_api_v0_prompts__delete",
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          },
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_NoneType_"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/not-found",
                  "title": "Prompt Not Found",
                  "detail": "Prompt with ID 'xxx' not found or access denied"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "type": "/errors/internal-server-error",
                  "title": "Internal Server Error",
                  "detail": "Failed to delete prompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromptData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the prompt",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Prompt name",
            "default": "Untitled Prompt"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Prompt description",
            "default": ""
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "LLM provider (e.g., openai, anthropic)",
            "default": "openai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model name (e.g., gpt-4, claude-3)",
            "default": "gpt-4"
          },
          "failover_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failover Model",
            "description": "Backup model if primary fails"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "Combined prompt content",
            "default": ""
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice",
            "description": "Tool choice configuration"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "description": "Sampling temperature",
            "default": 0.0
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P",
            "description": "Top-p sampling parameter"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "Maximum tokens to generate"
          },
          "response_format": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Format",
            "description": "Response format configuration"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream",
            "description": "Whether to stream the response"
          },
          "n_completions": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "N Completions",
            "description": "Number of completions to generate"
          },
          "stop": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop",
            "description": "Stop sequences"
          },
          "presence_penalty": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": -2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Presence Penalty",
            "description": "Presence penalty"
          },
          "frequency_penalty": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": -2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency Penalty",
            "description": "Frequency penalty"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed",
            "description": "Random seed for reproducibility"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API key override"
          },
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base",
            "description": "API base URL override"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User",
            "description": "User identifier for tracking"
          },
          "parallel_tool_calls": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parallel Tool Calls",
            "description": "Enable parallel tool calls"
          },
          "logprobs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logprobs",
            "description": "Return log probabilities"
          },
          "top_logprobs": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Logprobs",
            "description": "Number of top log probabilities"
          },
          "logit_bias": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logit Bias",
            "description": "Logit bias adjustments"
          },
          "stream_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Options",
            "description": "Streaming options"
          },
          "max_completion_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Completion Tokens",
            "description": "Maximum completion tokens"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "minimal",
                  "low",
                  "medium",
                  "high",
                  "auto"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort",
            "description": "Reasoning effort level",
            "default": "auto"
          },
          "extra_args": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Args",
            "description": "Additional provider-specific arguments"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Tags for prompt categorization"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "title": "PromptData",
        "description": "Core prompt data model with all fields that get saved to YAML files.\nThis model represents the complete prompt configuration including LLM parameters."
      },
...
      "ResponseStatus": {
        "type": "string",
        "enum": [
          "success",
          "error",
          "partial"
        ],
        "title": "ResponseStatus",
        "description": "Standard response status values."
      },
...
      "StandardResponse_NoneType_": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "description": "Response status indicator",
            "default": "success"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "description": "HTTP status code",
            "default": 200
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "Response payload"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable message about the response"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Response metadata"
          }
        },
        "type": "object",
        "title": "StandardResponse[NoneType]",
        "example": {
          "data": {
            "id": 1,
            "name": "Example"
          },
          "message": "Operation completed successfully",
          "meta": {
            "request_id": "req_123",
            "timestamp": "2024-01-01T00:00:00Z",
            "version": "1.0.0"
          },
          "status": "success"
        }
      },
      "StandardResponse_PromptMeta_": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "description": "Response status indicator",
            "default": "success"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "description": "HTTP status code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptMeta"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable message about the response"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Response metadata"
          }
        },
        "type": "object",
        "title": "StandardResponse[PromptMeta]",
        "example": {
          "data": {
            "id": 1,
            "name": "Example"
          },
          "message": "Operation completed successfully",
          "meta": {
            "request_id": "req_123",
            "timestamp": "2024-01-01T00:00:00Z",
            "version": "1.0.0"
          },
          "status": "success"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}